sanitize balances
