+++ src/handlers/updateBalances.ts
@@ -7,6 +7,7 @@ import { apiGatewayManagementApi } from '@handlers/apiGateway'
 import type { AdapterBalancesResponse } from '@handlers/getBalances'
 import { badRequest, serverError, success } from '@handlers/response'
 import { BaseContext, Contract } from '@lib/adapter'
+import { sanitizeBalances } from '@lib/balance'
 import { isHex, strToBuf } from '@lib/buf'
 import { getPricedBalances } from '@lib/price'
 import { isNotNullish } from '@lib/type'
@@ -132,33 +133,7 @@ export const websocketUpdateAdaptersHandler: APIGatewayProxyHandler = async (eve
     // Ungroup balances to make only 1 call to the price API
     const balances = adaptersBalances.flat().filter(isNotNullish)
 
-    // Filter out balances with invalid amounts
-    const sanitizedBalances = balances.filter((balance) => {
-      if (!balance.amount) {
-        console.error(`Missing balance amount`, balance)
-        return false
-      }
-
-      if (balance.underlyings) {
-        for (const underlying of balance.underlyings) {
-          if (!underlying.amount) {
-            console.error(`Missing underlying balance amount`, balance)
-            return false
-          }
-        }
-      }
-
-      if (balance.rewards) {
-        for (const reward of balance.rewards) {
-          if (!reward.amount) {
-            console.error(`Missing reward balance amount`, balance)
-            return false
-          }
-        }
-      }
-
-      return true
-    })
+    const sanitizedBalances = sanitizeBalances(balances)
 
     const hrstart = process.hrtime()
